chore(tests): address review cleanup items
