+++ platform/e2e-tests/utils/llm-provider-api-keys.ts
@@ -1,6 +1,10 @@
-import type { Page } from "@playwright/test";
+import type { APIRequestContext, Page } from "@playwright/test";
 import { E2eTestId } from "@shared";
-import { LLM_PROVIDER_API_KEYS_ROUTE } from "../consts";
+import {
+  API_BASE_URL,
+  LLM_PROVIDER_API_KEYS_ROUTE,
+  UI_BASE_URL,
+} from "../consts";
 import { expect, goToPage } from "../fixtures";
 import { clickButton, expandTablePagination } from "./dialogs";
 
@@ -88,6 +92,46 @@ export async function createVirtualKey(
   ).toBeVisible({
     timeout: 10_000,
   });
+}
+
+export async function deleteVisibleProviderKeys(
+  request: APIRequestContext,
+  provider: string,
+): Promise<void> {
+  const listResponse = await request.get(
+    `${API_BASE_URL}/api/llm-provider-api-keys?provider=${encodeURIComponent(provider)}`,
+    {
+      headers: {
+        "Content-Type": "application/json",
+        Origin: UI_BASE_URL,
+      },
+    },
+  );
+
+  if (!listResponse.ok()) {
+    throw new Error(
+      `Failed to list LLM provider API keys for ${provider}: ${listResponse.status()} ${await listResponse.text()}`,
+    );
+  }
+
+  const keys = (await listResponse.json()) as Array<{ id: string }>;
+  for (const key of keys) {
+    const deleteResponse = await request.delete(
+      `${API_BASE_URL}/api/llm-provider-api-keys/${key.id}`,
+      {
+        headers: {
+          "Content-Type": "application/json",
+          Origin: UI_BASE_URL,
+        },
+      },
+    );
+
+    if (!deleteResponse.ok()) {
+      throw new Error(
+        `Failed to delete LLM provider API key ${key.id}: ${deleteResponse.status()} ${await deleteResponse.text()}`,
+      );
+    }
+  }
 }
 
 async function getParentKeyOptionNameForProvider(
